Upgrade NDK to the version 9d
We can do upgrade of NDK up to the 9d with minimal changes. It brings users more modern
available platforms from android-9 up to android-19. Before they could choose
only android-14 as maximum platform.

The update of the default target version is proposed from android-14 to android-16
as the latest creates "position independent executables (PIE)" by default. It's not possible
to run not PIE on platforms >= android-16. It seems to be more logical to switch default
to support modern android android-16 to android-25 than supporting the previous versions.

It's still possible for user to compile to older versions, down to android-9. He just has
to manually switch the target version.